// src/components/form/entry/interface.ts
export interface PropsSelect {
  value?: any;
  showSearch?: boolean;
  maxTagCount?: number | 'responsive';
  onChange: (e: any) => any;
  onBlur?: (e: any) => any;
  placeholder?: string;
  className?: string;
  disabled?: boolean;
  get?: ITableGet;
  list?: ITableItemFilterList[];
  isMultiple?: boolean;
  allowClear?: boolean;
}